Revert "add function to download"
This reverts commit 4ad691c4595b911b5e4240b723f3e040cbc959d7.
