+++ crunchy-xml-decoder/altfuncs.py
@@ -35,8 +35,7 @@ def config():
     global localizecookies
     localizecookies = configr.getboolean('SETTINGS', 'localizecookies')
     onlymainsub = configr.getboolean('SETTINGS', 'onlymainsub')
-    connection_n_ = int(configr.get('SETTINGS', 'connection_n_'))
-    return [lang, lang2, forcesub, forceusa, localizecookies, quality, onlymainsub, connection_n_]
+    return [lang, lang2, forcesub, forceusa, localizecookies, quality, onlymainsub]
 
 
 #def playerrev(url):
@@ -110,11 +109,11 @@ def getxml(req, med_id):
             except:
                 try:
                     session.cookies['sess_id'] = re.split('"',requests.get('https://cr.onestay.moe/getid').text)[5]
-                    #print 'I recommend to re-login so we don\'t overload crunchyroll unblocker'
+                    print 'I recommend to re-login so we don\'t overload crunchyroll unblocker'
                 except:
                     sleep(10)  # sleep so we don't overload crunblocker
                     session.cookies['sess_id'] = re.split('"',requests.get('https://rubbix.net/crunchyroll/').text)[5]
-                    #print 'I recommend to re-login so we don\'t overload crunchyroll unblocker'
+                    print 'I recommend to re-login so we don\'t overload crunchyroll unblocker'
     headers = {'Referer': 'http://static.ak.crunchyroll.com/versioned_assets/ChromelessPlayerApp.17821a0e.swf',
                'Host': 'www.crunchyroll.com', 'Content-type': 'application/x-www-form-urlencoded',
                'User-Agent': 'Mozilla/5.0 (Windows NT 6.1; rv:26.0) Gecko/20100101 Firefox/26.0)'}
